a lot of bugfixes to get it to compile again

+++ branches/SVS2/Core/SVS/src/models/splinter.h
@@ -3,32 +3,32 @@
 
 #include "linalg.h"
 
-const double wheel_diameter     = 0.25;
-const double baseline           = 0.35;
-const double torque_constant    = 3.0; // torque (Nm) per amp
-const double emf_constant       = 2.0; // volts per rad_per_sec
-const double winding_resistance = 5.5; // ohms
-const double inertia            = 0.5; // kg*m^2
-const double drag_constant      = 1.0; // drag (Nm per rad_per_sec) ( >= 0)
-const double dt                 = 0.016; // need a better way to figure this out
+const float wheel_diameter     = 0.25;
+const float baseline           = 0.35;
+const float torque_constant    = 3.0; // torque (Nm) per amp
+const float emf_constant       = 2.0; // volts per rad_per_sec
+const float winding_resistance = 5.5; // ohms
+const float inertia            = 0.5; // kg*m^2
+const float drag_constant      = 1.0; // drag (Nm per rad_per_sec) ( >= 0)
+const float dt                 = 0.016; // need a better way to figure this out
 
-inline double calc_rps(double &rps, double input_volts) {
-	double volts_emf = rps * emf_constant;
-	double amps = (input_volts - volts_emf) / winding_resistance;
-	double torque0 = amps * torque_constant;
-	double torque_drag = rps * drag_constant;
-	double torque_net = torque0 - torque_drag;
-	double acceleration = torque_net / inertia;
+inline float calc_rps(float &rps, float input_volts) {
+	float volts_emf = rps * emf_constant;
+	float amps = (input_volts - volts_emf) / winding_resistance;
+	float torque0 = amps * torque_constant;
+	float torque_drag = rps * drag_constant;
+	float torque_net = torque0 - torque_drag;
+	float acceleration = torque_net / inertia;
 	rps += acceleration * dt;
 }
 
-inline void splinter_update(double lvolt, double rvolt, double &lrps, double &rrps, 
+inline void splinter_update(float lvolt, float rvolt, float &lrps, float &rrps, 
                      vec3 &pos, vec3 &vel, vec3 &rot, vec3 &rotrate)
 {
 	calc_rps(lrps, lvolt * 12);
 	calc_rps(rrps, rvolt * 12);
-	double dleft  = dt * lrps * wheel_diameter;
-	double dright = dt * rrps * wheel_diameter;
+	float dleft  = dt * lrps * wheel_diameter;
+	float dright = dt * rrps * wheel_diameter;
 	
 	quaternion orient(rot);
 	vel = orient.rotate(vec3((dleft + dright) / 2, 0, 0));
